+++ update_project.sh
@@ -7,7 +7,7 @@ BACKUP_FILE="RTSP Rotator.xcodeproj/project.pbxproj.backup"
 cp "$PROJECT_FILE" "$BACKUP_FILE"
 
 # Change product type from screensaver to application
-sed -i '' 's/PRODUCT_BUNDLE_IDENTIFIER = DisneyGPT.RTSP-Rotator;/PRODUCT_BUNDLE_IDENTIFIER = com.disneyg pt.rtsp-rotator;/g' "$PROJECT_FILE"
+sed -i '' 's/PRODUCT_BUNDLE_IDENTIFIER = com.jkoch.RTSP-Rotator;/PRODUCT_BUNDLE_IDENTIFIER = com.jkoch.rtsp-rotator;/g' "$PROJECT_FILE"
 sed -i '' 's/WRAPPER_EXTENSION = saver;/WRAPPER_EXTENSION = app;/g' "$PROJECT_FILE"
 sed -i '' 's/com.apple.product-type.bundle.screen-saver/com.apple.product-type.application/g' "$PROJECT_FILE"
 sed -i '' 's/productType = "com.apple.product-type.bundle.screen-saver"/productType = "com.apple.product-type.application"/g' "$PROJECT_FILE"